moves GameForm to own route and lifts related state
