+++ 4_client_server_db_jwt/client/src/App.jsx
@@ -3,46 +3,64 @@ import { useNavigate } from "react-router-dom";
 import Header from "./components/Header";
 import Dashboard from "./components/Dashboard";
 import GameDetail from "./components/GameDetail";
-import { useAuth } from './providers/authProvider'
+import { useAuth } from "./providers/authProvider";
 import { authenticateFetch } from "./api";
-import Routes from './routes/Routes'
+import Routes from "./routes/Routes";
 import GridLoader from "react-spinners/GridLoader";
 
-
-
 function App() {
-
   const { token, setToken } = useAuth();
   const [isLoading, setIsLoading] = useState(true);
-  
+  const [games, setGames] = useState([]);
 
- useEffect(() => {
+  useEffect(() => {
     setIsLoading(true);
-   const fetchAuth = async () => {
-     const res = await authenticateFetch();
-     if (res.ok) {
-       const tokenJSON = await res.json();
-       setToken(tokenJSON);
-     } else {
-       setToken(null);
-     }
-     setIsLoading(false);
-   };
- 
-   fetchAuth();
- }, []);
+    const fetchAuth = async () => {
+      const res = await authenticateFetch();
+      if (res.ok) {
+        const tokenJSON = await res.json();
+        setToken(tokenJSON);
+      } else {
+        setToken(null);
+      }
+      setIsLoading(false);
+    };
 
- if (isLoading) return <GridLoader />;
+    fetchAuth();
+  }, []);
+
+  useEffect(() => {
+    const fetchGames = async () => {
+      const response = await fetch("/games");
+      const gameArr = await response.json();
+      setGames(gameArr);
+    };
+    fetchGames().catch(console.error);
+  }, []);
+
+  function handleAddGame(newGame) {
+    setGames((games) => [...games, newGame]);
+  }
+
+  function handleDeleteGame(id) {
+    fetch(`/games/${id}`, { method: "DELETE" }).then((r) => {
+      if (r.ok) {
+        setGames((games) => games.filter((games) => games.id !== id));
+      }
+    });
+  }
+
+  if (isLoading) return <GridLoader />;
 
   return (
     <div>
-       
-        <main>
-          
-            <Routes />
-         
-        </main>
-        
+      <main>
+        <Routes
+          games={games}
+          handleAddGame={handleAddGame}
+          handleDeleteGame={handleDeleteGame}
+        />
+      </main>
     </div>
   );
 }
