+++ 4_client_server_db_jwt/client/src/components/Dashboard.jsx
@@ -3,29 +3,8 @@ import GameCard from "./GameCard";
 import GameForm from "./GameForm";
 import GridLoader from "react-spinners/GridLoader";
 
-function Dashboard() {
-  const [games, setGames] = useState([]);
-
-  useEffect(() => {
-    const fetchGames = async () => {
-      const response = await fetch("/games");
-      const gameArr = await response.json();
-      setGames(gameArr);
-    };
-    fetchGames().catch(console.error);
-  }, []);
-
-  function handleAddGame(newGame) {
-    setGames((games) => [...games, newGame]);
-  }
-
-  function handleDeleteGame(id) {
-    fetch(`/games/${id}`, { method: "DELETE" }).then((r) => {
-      if (r.ok) {
-        setGames((games) => games.filter((games) => games.id !== id));
-      }
-    });
-  }
+function Dashboard({games, handleDeleteGame}) {
+  
 
   let gameCards = games.map((game) => (
     <GameCard key={game.id} game={game} onDelete={handleDeleteGame} />
@@ -37,8 +16,6 @@ function Dashboard() {
         <h1>Your Games</h1>
         <div className="gameList">{gameCards}</div>
       </Suspense>
-      <hr />
-      <GameForm onGameRequest={handleAddGame} />
     </>
   );
 }
